Testing with no deadzone for G29 and new steering limits
Before this commit, there was a deadzone in the joystick ROS node that
prevented steering commands from being sent before the joystick (or steering
wheel) had surpassed a certain percentage of the overall range of movement.
After this commit, there is no deadzone, meaning the full range of the
G29 is capable of sending steering commands. This deadzone is still useful for
joystick based controllers. Additionally, there are new values for the steering
range for Vehicle 2.